# src/swagger.json
{
    "openapi":"3.0.0",
    "info":{
        "title":"API de cadastro de usuários",
        "description":"Essa api tem como principal objetivo o cadastro de novos usuários",
        "contact":{
            "email":"josecatanaoneto@gmail.com"
        },
        "version":"1.0.0"
    },
    "servers":[
        {
            "url":"localhost:3000/",
            "description":"API de teste"
        }
    ],
    "paths":{
        "/login":{
            "post":{
                "summary":"Faz o login do usuário",
                "description":"Essa rota será responsavel por fazer o login do usuário",
                "tags":["Users"],
                "requestBody":{
                    "content":{
                        "application/json":{
                            "schema":{
                                "properties":{
                                    "name":{
                                        "type":"string"
                                    },
                                    "password":{
                                        "type":"string"
                                    } 
                                }
                            },
                            "examples":{
                                "User":{
                                    "value":{
                                        "name":"jose",
                                        "password":"123"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/cadastro":{
            "post":{
                "summary":"Cadastra o usuario.",
                "description":"Essa rota será responsavel por cadastrar usuário",
                "tags":["Users"],
                "requestBody":{
                    "content":{
                        "application/json":{
                            "schema":{
                                "properties":{
                                    "name":{
                                        "type":"string"
                                    },
                                    "email":{
                                        "type":"string"
                                    },
                                    "password":{
                                        "type":"string"
                                    } 
                                }
                            },
                            "examples":{
                                "User":{
                                    "value":{
                                        "name":"jose",
                                        "email":"josecatanoneto@gmail.com",
                                        "password":"123"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}